+++ lib/palettes.dart
@@ -1,0 +1,4 @@
+/// The color palettes.
+
+export 'src/palette.dart';
+export 'src/palette_rgb.dart';